Allow to remove a word from developer dictionary
Remove words from all known dictionaries, local and Git.

Fixes T2136.

Test Plan:
  - `bin/merge-dictionaries --delete alowed`

Differential Revision: https://devcentral.nasqueron.org/D3724

+++ src/mergedictionaries/write/git.py
@@ -6,10 +6,14 @@
 #                   from Git repository
 #   License:        BSD-2-Clause
 #   -------------------------------------------------------------
+
+
 import os
 from tempfile import NamedTemporaryFile
+from typing import List
 
 from mergedictionaries.sources import GitRepository
+from mergedictionaries.utils.collections import remove_words
 
 
 def build_temporary_dictionary(words):
@@ -29,5 +33,20 @@ def write(words, target_repos, cached_repos):
 
     for repo in target_repos:
         GitRepository(repo, cached_repos).publish(tmp_dictionary_path)
+
+    os.unlink(tmp_dictionary_path)
+
+
+def delete_words(repo: GitRepository, words_to_delete: List):
+    current_words = repo.extract_words()
+
+    if not any(word in current_words for word in words_to_delete):
+        # Nothing to do, the dictionary is already up to date.
+        return
+
+    words = remove_words(current_words, words_to_delete)
+
+    tmp_dictionary_path = build_temporary_dictionary(words)
+    repo.publish(tmp_dictionary_path)
 
     os.unlink(tmp_dictionary_path)
